Ready http wrapper for documentation
  - Make public API file consist of one exported function
  - Move required sub-functions to utils file in http folder
  - Reconfigure tests accordingly
