+++ lib/transports/http.js
@@ -15,30 +15,15 @@
  */
 
 'use strict';
-var https = require('https');
-var http = require('http');
-var isFunction = require('lodash.isfunction');
-var isString = require('lodash.isstring');
-var has = require('lodash.has');
-var isObject = require('lodash.isobject');
-var noOp = require('lodash.noop');
 
-var utils = {
-  url: require('./http/url'),
-  headers: require('./http/headers'),
-  payload: require('./http/payload')
-};
-
-function getTransport (isUsingHttps) {
-  if (isUsingHttps) {
-    return https;
-  }
-  return http;
-}
+var url = require('./http/url');
+var headers = require('./http/headers');
+var payload = require('./http/payload');
+var utils = require('./http/utils');
 
 function formulateRequest (generatedOptions, cb) {
   var body = '';
-  var request = getTransport(generatedOptions.isUsingHttps).request(
+  var request = utils.getTransport(generatedOptions.isUsingHttps).request(
     generatedOptions.requestOptions, function (response) {
       response.on('data', function (chunk) {
         body += chunk;
@@ -58,76 +43,19 @@ function formulateRequest (generatedOptions, cb) {
   return request;
 }
 
-function generateErrorInformation (body, res) {
-  var err = null;
-  if (isObject(body) && has(body, 'error') && res.statusCode !== 200) {
-    if (isString(body.error)) {
-      err = new Error(body.error);
-      err.code = res.statusCode;
-      return err;
-    } else if (Array.isArray(body.error.errors)) {
-      err = new Error(body.error.errors.map(
-        function (err) {
-          return err.message;
-        }
-      ).join('\n'));
-      err.code = body.error.code;
-      err.errors = body.error.errors;
-      return err;
-    }
-    err = new Error(body.error.message);
-    err.code = has(body.error, 'code') ? body.error.code : res.statusCode;
-  } else if (res.statusCode >= 500) {
-    err = new Error(body);
-    err.code = res.statusCode;
-  }
-  return err;
-}
-
-function attemptBodyParse (body) {
-  try {
-    return JSON.parse(body);
-  } catch (e) {
-    return body;
-  }
-}
-
-function wrapGivenCallback (cb) {
-  if (!isFunction(cb)) {
-    return noOp;
-  }
-  return function (err, res, body) {
-    var parsedBody, parsedError;
-    if (err || !body) {
-      return cb(err, body, res);
-    }
-    parsedBody = attemptBodyParse(body);
-    parsedError = generateErrorInformation(parsedBody, res);
-    if (parsedError) {
-      parsedBody = null;
-    }
-    cb(parsedError, parsedBody, res);
-  };
-}
-
-function normalizeRequestOptions (givenOptions) {
-  return utils.url.generateUrl(utils.headers.generateHeaders(
-    utils.payload.generateRequestPayload({}, givenOptions), givenOptions),
-    givenOptions
-  );
-}
-
 function request (givenOptions, cb) {
   return formulateRequest(
-    normalizeRequestOptions(givenOptions),
-    wrapGivenCallback(cb)
+    url.generateUrl(
+      headers.generateHeaders(
+        payload.generateRequestPayload({}, givenOptions),
+        givenOptions
+      ),
+      givenOptions
+    ),
+    utils.wrapGivenCallback(cb)
   );
 }
 
 module.exports = {
-  request: request,
-  getTransport: getTransport,
-  wrapGivenCallback: wrapGivenCallback,
-  generateErrorInformation: generateErrorInformation,
-  normalizeRequestOptions: normalizeRequestOptions
+  request: request
 };
